Add LayoutToken Table
